Issue #121: Moved ColoredCoinContext out of Wallet Model
* moved ColoredCoinContext from wallet_model.py to color.py
* added unit tests for color.py
* refactored ColoredCoinContext a little bit
